+++ designer/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/component/ComponentApiEndpoints.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.restmodel.component
 
 import cats.data.NonEmptySet
-import pl.touk.nussknacker.engine.api.component.ProcessingMode.AllowedProcessingModes
+import pl.touk.nussknacker.engine.api.component.Component.AllowedProcessingModes
 import pl.touk.nussknacker.engine.api.component.{
   ComponentGroupName,
   ComponentType,
@@ -66,8 +66,7 @@ class ComponentApiEndpoints(auth: EndpointInput[AuthCredentials]) extends BaseEn
                       )
                     ),
                     usageCount = 2,
-                    allowedProcessingModes =
-                      AllowedProcessingModes.SetOf(NonEmptySet.one(ProcessingMode.RequestResponse))
+                    allowedProcessingModes = AllowedProcessingModes.SetOf(ProcessingMode.RequestResponse)
                   )
                 )
               )